move import/export tests to their own class
